% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{agg_co_units}
\alias{agg_co_units}
\title{aggregate sc weights to original tx and (weighted) co units}
\usage{
agg_co_units(scweights)
}
\arguments{
\item{scweights}{list of sc weights (tibbles with tx unit in first row)}
}
\value{
df with original tx and sc units, with weights
}
\description{
aggregate sc weights to original tx and (weighted) co units
}
