% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{agg_sc_units}
\alias{agg_sc_units}
\title{aggregate scweights to tx/sc units}
\usage{
agg_sc_units(scweights)
}
\arguments{
\item{scweights}{list of sc weights (tibbles with tx unit in first row)}
}
\value{
df with tx and corresponding sc units
}
\description{
aggregate scweights to tx/sc units
}
