% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{boot_bayesian}
\alias{boot_bayesian}
\title{Run a Bayesian bootstrap}
\usage{
boot_bayesian(d, B = 100)
}
\arguments{
\item{d}{dataset with weights for each tx/co unit}

\item{B}{number of bootstrap samples}
}
\value{
vector of bootstrapped ATT estimates
}
\description{
Run a Bayesian bootstrap
}
