% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{boot_bayesian2}
\alias{boot_bayesian2}
\title{Run a Bayesian bootstrap, corrected (?)}
\usage{
boot_bayesian2(d, B = 100)
}
\arguments{
\item{d}{dataset with tx units and sc units}

\item{B}{number of bootstrap samples}
}
\value{
vector of bootstrapped ATT estimates
}
\description{
Run a Bayesian bootstrap, corrected (?)
}
