% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{boot_bayesian_redux}
\alias{boot_bayesian_redux}
\title{#' @param d dataset with tx units and sc units
#' @param B number of bootstrap samples
#'
#' @return vector of bootstrapped ATT estimates
boot_bayesian2_wild <- function(d, B=100) {
map_dbl(1:B,
.progress = "Bootstrapping...",
function(b) {
d \%>\%
mutate(
boot_weights = rep(
sample(c(-(sqrt(5)-1)/2, (sqrt(5)+1)/2),
prob = c((sqrt(5)+1)/(2\emph{sqrt(5)), (sqrt(5)-1)/(2}sqrt(5))),
size=nrow(d)/2, replace=T),
each=2),
weights = weights * boot_weights) \%>\%
get_att_ests()
})
}}
\usage{
boot_bayesian_redux(d, B = 100)
}
\description{
#' @param d dataset with tx units and sc units
#' @param B number of bootstrap samples
#'
#' @return vector of bootstrapped ATT estimates
boot_bayesian2_wild <- function(d, B=100) {
map_dbl(1:B,
.progress = "Bootstrapping...",
function(b) {
d \%>\%
mutate(
boot_weights = rep(
sample(c(-(sqrt(5)-1)/2, (sqrt(5)+1)/2),
prob = c((sqrt(5)+1)/(2\emph{sqrt(5)), (sqrt(5)-1)/(2}sqrt(5))),
size=nrow(d)/2, replace=T),
each=2),
weights = weights * boot_weights) \%>\%
get_att_ests()
})
}
}
