% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot_by_resids}
\alias{boot_by_resids}
\title{Title}
\usage{
boot_by_resids(resids, B, boot_mtd, seed_addition = 123)
}
\arguments{
\item{resids}{(desc)}

\item{B}{(desc)}

\item{boot_mtd}{(desc)}

\item{seed_addition}{(desc)}
}
\value{
(desc)
}
\description{
Title
}
