% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot_CSM}
\alias{boot_CSM}
\title{Title}
\usage{
boot_CSM(
  dgp_name,
  att0,
  I = 100,
  B = 250,
  mu_model = "linear",
  boot_mtd = "Bayesian",
  n_split = 1,
  kang_true = FALSE,
  toy_ctr_dist = 0.5
)
}
\arguments{
\item{dgp_name}{Name of the DGP}

\item{att0}{True ATT}

\item{I}{Number of MC runs}

\item{B}{Number of bootstrap samples per MC run}

\item{mu_model}{Model for the potential outcomes}

\item{boot_mtd}{Method for the bootstrap}

\item{n_split}{Number of splits for sample splitting}

\item{kang_true}{Whether to use the true ATT from Kang et al. (2021)}
}
\description{
Title
}
\examples{
boot_otsu_to_test <-
boot_CSM(dgp_name="otsu", att0=T,I=100, B=100,mu_model="linear", boot_mtd="Bayesian",n_split=2)
}
