% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{boot_naive}
\alias{boot_naive}
\title{Run naive bootstrap}
\usage{
boot_naive(
  d,
  caliper,
  metric,
  rad_method,
  dist_scaling,
  est_method = "scm",
  knn = 25,
  B = 50
)
}
\arguments{
\item{d}{original dataset}

\item{B}{number of bootstrap samples}
}
\value{
vector of bootstrapped ATT estimates
}
\description{
Run naive bootstrap
}
