% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot_SE}
\alias{boot_SE}
\title{Estimate the variance from the bootstrap}
\usage{
boot_SE(
  matches_table,
  outcome = "Y",
  treatment = "Z",
  B = 100,
  boot_mtd = "sign"
)
}
\arguments{
\item{matches_table}{The data frame of the matched table}

\item{outcome}{Name of the outcome variable (default "Y")}

\item{treatment}{Name of the treatment variable (default "Z")}

\item{var_weight_type}{The way that cluster variances are averaged
"num_units": weight by number of units in the subclass
"ess_units": weight effective size of units in the subclass
"uniform: weight each cluster equally.}
}
\value{
A tibble with SE, sigma_hat, N_T, and N_C_tilde
}
\description{
This is the main bootstrap function.
}
