% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_overlap_stat.R
\name{calculate_overlap_statistics_from_match_object}
\alias{calculate_overlap_statistics_from_match_object}
\title{Calculate Overlap Statistics from Match Object}
\usage{
calculate_overlap_statistics_from_match_object(mtch)
}
\arguments{
\item{mtch}{A match object from matching procedures.}
}
\value{
A list containing all three sets of overlap statistics.
}
\description{
Extracts a matched table from a match object and then calculates all three
overlap metrics.
}
