% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_overlap_stat.R
\name{calculate_overlap_stats_from_table}
\alias{calculate_overlap_stats_from_table}
\title{Calculate Overlap Statistics from Matched Table}
\usage{
calculate_overlap_stats_from_table(full_matched_table)
}
\arguments{
\item{full_matched_table}{A data frame containing the matched data.}
}
\value{
A list containing all three sets of overlap statistics.
}
\description{
Calculates all three overlap metrics for a matched dataset from a
pre-existing full matched table.
}
