% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{calculate_subclass_variances}
\alias{calculate_subclass_variances}
\title{Calculate the variance for each subclass}
\usage{
calculate_subclass_variances(matches_filtered, outcome = "Y")
}
\arguments{
\item{matches_filtered}{The filtered matched object}

\item{outcome}{Name of the outcome variable (default "Y")}
}
\value{
A data frame with subclass variances
}
\description{
Calculate the variance for each subclass
}
