% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{calculate_weighted_variance}
\alias{calculate_weighted_variance}
\title{Calculate the weighted average variance
NOTE: Could weight by nj, w_nj, or 1.  w_nj takes into account how}
\usage{
calculate_weighted_variance(cluster_var_df, var_weight_type = "num_units")
}
\arguments{
\item{cluster_var_df}{Data frame. Each row is a subclass and
its variances, its weight}

\item{var_weight_type}{The way that cluster variances are averaged
"num_units": weight by number of units in the subclass
"ess_units": weight effective size of units in the subclass
"uniform: weight each cluster equally}
}
\value{
Weighted average of subclass variances
}
\description{
Calculate the weighted average variance
NOTE: Could weight by nj, w_nj, or 1.  w_nj takes into account how
}
