% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_plot.R
\name{caliper_sensitivity_table}
\alias{caliper_sensitivity_table}
\alias{caliper_sensitivity_plot}
\alias{caliper_sensitivity_plot_stats}
\title{Make sensitivity table or plot of impact of changing caliper}
\usage{
caliper_sensitivity_table(
  csm,
  data,
  outcome = "Y",
  min_cal = 0.05,
  max_cal = 5,
  R = 30
)

caliper_sensitivity_plot(
  csm,
  data,
  outcome = "Y",
  min_cal = 0.05,
  max_cal = 5,
  R = 30,
  focus = c("FATT", "ATT", "FATT_1nn", "ATT_1nn", "FATT_raw", "ATT_raw"),
  lines = c("FATT", "ATT", "FATT_1nn", "ATT_1nn", "FATT_raw", "ATT_raw")
)

caliper_sensitivity_plot_stats(
  csm,
  data,
  outcome = "Y",
  min_cal = 0.05,
  max_cal = 5,
  R = 30,
  focus = c("FATT", "ATT", "FATT_1nn", "ATT_1nn", "FATT_raw", "ATT_raw"),
  lines = c("FATT", "ATT", "FATT_1nn", "ATT_1nn", "FATT_raw", "ATT_raw")
)
}
\arguments{
\item{csm}{A CSM object OR a sensitivity ggplot created by, e.g.,
\code{\link[=caliper_sensitivity_plot]{caliper_sensitivity_plot()}} or
\code{\link[=caliper_sensitivity_plot_stats]{caliper_sensitivity_plot_stats()}} (or anything with the
sensitivity table as an attribute) OR a table with the relevant
columns (focus, lines, caliper, ATT, SE, ESS_C). This allows
calls to the plot methods without recalculating everything.}

\item{data}{The data the csm_matches object was fit to. (One could
put in alternate data here, if the covariates aligned.)}

\item{min_cal}{Minimum caliper to try}

\item{max_cal}{Maximum caliper to try}

\item{R}{Number of calipers to try between min and max}

\item{focus}{What type of estimate to plot the confidence envelope
for.}

\item{lines}{Which estimates to show as lines on the plot.}
}
\value{
Either a table of results, or a ggplot object of the
specified sensitivity plot.  All generated plots have an
attribute of "table" that holds the data used to make the plot,
and these can be fed back into these methods to change the plot.
}
\description{
To generate the sensitivity table, method will run caliper from a
min to max value, generating matches at each step. Then make a plot
of the ATT and confidence interval for each point on the plot. Also
can return table of results if 'return_table' is set to TRUE.
}
\details{
\code{caliper_sensitivity_table()} generates the sensitivity
table only.

\code{caliper_sensitivity_plot()} generates a ggplot with
y-axis as ATT estimate and confidence interval, and x-axis a series
of calipers from the minimum to maximum specified.  Different lines
are shown for different ATT estimates.

\code{caliper_sensitivity_plot_stats()} makes an augmented plot
from sensitivity plot showing ATT, SE and ESS
}
