% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess_tools.R
\name{compare_ess}
\alias{compare_ess}
\title{Compare Effective Sample Size (ESS) across methods}
\usage{
compare_ess(csm, data)
}
\arguments{
\item{csm}{A csm_matches object}

\item{data}{The data the csm_matches object was fit to.}
}
\value{
A tibble comparing the ESS for each method
}
\description{
Given a csm_matches object, this function computes and compares
the Effective Sample Size (ESS) for three different estimation
methods: Synthetic Control Method (SCM), Average, and 1-Nearest
Neighbor (1-NN).
}
