% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_overlap_stat.R
\name{compute_pairwise_overlap_statistics}
\alias{compute_pairwise_overlap_statistics}
\title{Compute Pairwise Overlap Statistics}
\usage{
compute_pairwise_overlap_statistics(pairwise_shared_matrix)
}
\arguments{
\item{pairwise_shared_matrix}{A matrix from compute_pairwise_shared_controls}
}
\value{
A list with median, 75th percentile, and max statistics
}
\description{
Computes summary statistics for pairwise overlap of shared controls.
This represents how many controls are shared between pairs of treated units.
}
