% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_matches_object.R
\name{is.csm_matches}
\alias{is.csm_matches}
\alias{[.csm_matches}
\alias{[[.csm_matches}
\alias{dim.csm_matches}
\alias{as.data.frame.csm_matches}
\title{Dimension of csm_matches object}
\usage{
is.csm_matches(x)

\method{[}{csm_matches}(x, ...)

\method{[[}{csm_matches}(x, ...)

\method{dim}{csm_matches}(x, ...)

\method{as.data.frame}{csm_matches}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{...}{additional arguments to be passed to the
as.data.frame.list methods.}

\item{row.names}{NULL or a character vector giving the
row names for the data frame.}

\item{optional}{logical. If TRUE, setting row names and
converting column names is optional.}
}
\value{
is.csm_matches: TRUE if object is a csm_matches object.

\code{[}: pull out rows and columns of the dataframe.

\code{[[}: pull out single element of dataframe.

dim: Dimension of csm_matches (as matrix)

as.data.frame: csm_matches object as a clean
dataframe (no more attributes from csm_matches).
}
\description{
Dimension of csm_matches object

Cast csm_matches result to data.frame
}
