% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{ess}
\alias{ess}
\title{Effective Sample Size}
\usage{
ess(weights)
}
\arguments{
\item{weights}{A numeric vector of weights}
}
\value{
The effective sample size
}
\description{
Calculate the effective sample size given a vector of unit weights
}
