% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{est_weights}
\alias{est_weights}
\title{Estimate control weights within matched sets}
\usage{
est_weights(
  matched_gps,
  est_method = c("scm", "average"),
  covs = params(matched_gps)$covariates,
  treatment = params(matched_gps)$treatment,
  scaling = params(matched_gps)$scaling,
  metric = params(matched_gps)$metric
)
}
\arguments{
\item{matched_gps}{Either a csm_matches object or a list of matched
sets, with a tx unit in first row of each set.}

\item{est_method}{"scm" or "average"}

\item{covs}{List of covariates to calculate similarity scores on if
using scm.}

\item{scaling}{tibble with scaling for each covariate}
}
\value{
If passed an csm object, an updated csm object with
weights. Otherwise, return list of matched sets, with 'weights'
for each control unit.
}
\description{
Generate weights for list of matched sets, typically by the
synthetic control method (but you can also simply average).
}
