% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_plot.R
\name{feasible_plot}
\alias{feasible_plot}
\title{Make feasible plot showing cumulative ATT as feasible units are
added}
\usage{
feasible_plot(csm, return_table = FALSE, caliper_plot = FALSE)
}
\arguments{
\item{csm}{A csm_matches object}

\item{return_table}{If TRUE, return the full table of results
instead of a plot}

\item{caliper_plot}{If TRUE, return a plot of maximum caliper size
vs number of treated units.  If "both" return a list of both
plots, along with the table itself.}
}
\value{
A ggplot object showing the feasible plot, or a table of
results if return_table is TRUE.
}
\description{
Given a csm_matches object, this function makes a plot showing how
the cumulative ATT estimate changes as more feasible treated units
are added (i.e., as the caliper size increases to include more
treated units).
}
