% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_matches_object.R
\name{feasible_unit_subclass}
\alias{feasible_unit_subclass}
\title{List all IDs of subclasses of units that are feasible}
\usage{
feasible_unit_subclass(csm)
}
\value{
Numeric vector of IDs
}
\description{
For all units matched without adapting the caliper, get the
subclass IDs (i.e., the ids that link controls to treated units in
matched clusters).
}
