% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_matches_object.R
\name{full_unit_table}
\alias{full_unit_table}
\title{Get full raw table of all the units}
\usage{
full_unit_table(csm, feasible_only = FALSE, nonzero_weight_only = FALSE)
}
\arguments{
\item{csm}{A csm_matches object from a matching call.}

\item{feasible_only}{TRUE means only return treated units and
matched controls for units that could be matched within the set
caliper.}

\item{nonzero_weight_only}{TRUE means drop any control units with 0
weight (e.g., due to scm weighting).}
}
\value{
dataframe
}
\description{
Return dataframe with control units repeated, grouped with treated
unit with individual weights with those treated units.  Treated
units also included.
}
