% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{gen_df_acic}
\alias{gen_df_acic}
\title{Generate data from the ACIC 2016 competition}
\usage{
gen_df_acic(
  model.trt = "step",
  root.trt = 0.35,
  overlap.trt = "full",
  model.rsp = "linear",
  alignment = 0.75,
  te.hetero = "high",
  random.seed = 1,
  n = 1000,
  p = 10
)
}
\value{
A tibble containing the generated dataset with covariates,
treatment assignment, potential outcomes, and observed outcome.
}
\description{
This function generates a dataset using the data generating process
from the ACIC 2016 competition. It samples covariates from the
input_2016 dataset and then applies the dgp_2016 function to
generate potential outcomes and treatment assignments.
}
