% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{gen_df_adv}
\alias{gen_df_adv}
\title{Generate toy data}
\usage{
gen_df_adv(
  nc,
  nt,
  f0_sd = 0.1,
  f0_fun = function(X1, X2) {
     1
 },
  tx_effect_fun = function(X1, X2) {
     1
 },
  f0_sd_fun = NULL,
  ctr_dist = 0.5,
  prop_nc_unif = 1/3
)
}
\arguments{
\item{nc}{(something)}

\item{nt}{(something)}

\item{f0_sd}{degree of residual noise (default homoskedastic)}

\item{f0_fun}{(something)}

\item{tx_effect_fun}{(something)}

\item{f0_sd_fun}{Function to generate noise standard deviation.
Default is NULL. If not null, override f0_sd.}

\item{ctr_dist}{from 0 to 1; lower means better overlap}

\item{prop_nc_unif}{proportion of uniform controls. lower means
worse overlap}
}
\value{
A tibble containing a toy dataset
}
\description{
generate data mostly in grid (0,1) x (0,1)
}
