% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{gen_df_kang}
\alias{gen_df_kang}
\title{Simulation data from the Kang paper}
\usage{
gen_df_kang(n = 1000)
}
\arguments{
\item{n}{Number of samples to generate}
}
\value{
A data frame with covariates and outcome
}
\description{
A DGP for simulation studies.
}
\details{
Generates a 4 dimensional multivariate normal that then gets
transformed to generate four "X" variables.  The Outcome is a
function of the original V variables.
}
