% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{gen_dm}
\alias{gen_dm}
\title{Generate (#tx) by (#co) distance matrix}
\usage{
gen_dm(
  data,
  covs = get_x_vars(data),
  treatment = "Z",
  scaling = 1,
  metric = c("maximum", "euclidean", "manhattan")
)
}
\arguments{
\item{data}{dataframe contains at least covariates and treatments
specified by "covs" and "treatment" argument}

\item{covs}{a vector of characters,
name of covariates to be matched; default all "X" variables}

\item{treatment}{name of the treatment variable; default "Z"}

\item{scaling}{vector of scaling constants for covariates}

\item{metric}{a character vector, distance metric to use.
Default is c("maximum", "euclidean", "manhattan")}
}
\value{
(#tx) by (#co) distance matrix
}
\description{
Generate (#tx) by (#co) distance matrix
}
