% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{gen_toy_covar}
\alias{gen_toy_covar}
\title{Generate covariates X1, X2 for the toy example}
\usage{
gen_toy_covar(n, X1_ctrs, X2_ctrs, SD)
}
\arguments{
\item{n}{Number of samples}

\item{X1_ctrs}{A vector of two elements, representing the two different
X1 locations}

\item{X2_ctrs}{Same as X1_ctrs, but}

\item{SD}{Std. deviation of the}
}
\value{
A tibble object with two columns X1, X2
}
\description{
Generate covariates X1, X2 for the toy example
}
\examples{
n = 100
X1_ctrs <- c(0.25, 0.75)
X2_ctrs <- c(0.25, 0.75)
SD <- 0.1
res <- gen_toy_covar(n, X1_ctrs, X2_ctrs, SD)
}
