% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{gen_toy_covar_k}
\alias{gen_toy_covar_k}
\title{Generate k-dimensional blobs around specified centers
(Helper function replacing the original 2D gen_toy_covar)}
\usage{
gen_toy_covar_k(n, centers, k, sd)
}
\arguments{
\item{n}{Number of points to generate.}

\item{centers}{A list of k-dimensional numeric vectors representing cluster centers.}

\item{k}{Dimensionality.}

\item{sd}{Standard deviation for the blobs (assumed spherical).}
}
\value{
A tibble with k columns (X1, ..., Xk).
}
\description{
Generate k-dimensional blobs around specified centers
(Helper function replacing the original 2D gen_toy_covar)
}
