% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_ATT}
\alias{estimate_ATT}
\title{Estimate ATT with correct variance estimation}
\usage{
estimate_ATT(
  scmatch,
  treatment = "Z",
  outcome = "Y",
  superpopulation = FALSE,
  var_weight_type = "ess_units",
  variance_method = c("pooled", "bootstrap"),
  boot_mtd = "wild",
  B = 250,
  seed_addition = 11,
  cluster_comb_mtd = "average",
  df = NULL,
  ...
)
}
\arguments{
\item{scmatch}{The CSM match object, an R S3 object}

\item{treatment}{Name of the treatment variable (default "Z")}

\item{outcome}{Name of the outcome variable (default "Y")}

\item{superpopulation}{TRUE means use superpopulation-targeting
variance estimator, FALSE means finite population targeting.}

\item{var_weight_type}{The way that cluster variances are averaged
(default "ess_units")}

\item{variance_method}{Method for calculating measurement error
variance: "pooled": use get_measurement_error_variance (default)
"bootstrap": use get_measurement_error_variance_OR}

\item{boot_mtd}{Bootstrap method when variance_method = "bootstrap"
(default: "wild")}

\item{B}{Number of bootstrap samples when variance_method =
"bootstrap" (default: 250)}

\item{seed_addition}{Additional seed for bootstrap (default: 11)}

\item{df}{The \emph{full} original data frame. Required for 'ai06'
method.}

\item{...}{Additional arguments passed to get_total_variance.}
}
\value{
A tibble with ATT estimate, standard error, t-statistic,
and variance components
}
\description{
Calculate ATT and associated standard error using the specified
variance estimator (depending on provided arguments).
}
