% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{get_att_ests}
\alias{get_att_ests}
\title{Estimate the ATT from a matched dataframe}
\usage{
get_att_ests(matched_df, treatment = "Z", outcome = "Y")
}
\arguments{
\item{matched_df}{A matched dataset}
}
\description{
Given a matched dataset, calculate the estimated ATT
}
