% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{get_att_point_est}
\alias{get_att_point_est}
\title{Estimate the ATT from a matched dataframe}
\usage{
get_att_point_est(matched_df, treatment = "Z", outcome = "Y")
}
\arguments{
\item{matched_df}{A matched dataset, either the dataframe of
treatment and control units, or a csm_matches object.  If
dataframe, needs the treatment, outcome, and a "weights" column.}
}
\description{
Given a matched dataset, calculate the estimated ATT
}
