% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{get_distance_table}
\alias{get_distance_table}
\title{Calculate distances from all matched treatment units to controls}
\usage{
get_distance_table(scm, long_table = FALSE)
}
\arguments{
\item{scm}{A csm_matches object}

\item{long_table}{If TRUE, return a long-form table with method and
distance columns.  If FALSE each tx unit is a row.}
}
\value{
A data frame with distances from each treated unit to their
synthetic control, average control, and closest control.
}
\description{
Look at distances between each treated unit and their synthetic
control, average control, and closest control.
}
