% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_matches_object.R
\name{get_match_sets}
\alias{get_match_sets}
\title{Return result table for listed treated units}
\usage{
get_match_sets(csm, id, nonzero_weight_only = TRUE)
}
\arguments{
\item{csm}{A csm_matches object}

\item{id}{list of ids for treated units}

\item{nonzero_weight_only}{TRUE means drop any control units with 0
weight (e.g., due to scm weighting).}
}
\value{
Dataframe of matches
}
\description{
Return result table for listed treated units
}
