% Generated by roxygen2: do not edit by hand
% Please edit documentation in scripts/wrappers.R
\name{get_matches}
\alias{get_matches}
\title{Perform Matching Based on Specified Type}
\usage{
get_matches(matching_type, df_dgp, scaling, k = 8)
}
\arguments{
\item{matching_type}{A string specifying the type of matching. Supported types are:
\itemize{
\item \code{"maximum_fixed_scm"}: Uses the maximum metric with fixed radius and SCM estimation.
\item \code{"euclidean_knn"}: Uses the Euclidean metric with k-nearest neighbors.
}}

\item{df_dgp}{A data frame containing the dataset to be matched. The dataset should include
treatment and covariate columns.}

\item{scaling}{A numeric value or vector used for scaling covariates during the matching process.
Defaults to \code{1}.}

\item{k}{A numeric value to specify the number of nearest neighbors in knn matching
Defaults to \code{8}}
}
\value{
A list containing the matched dataset and associated metrics. The specific structure of
the result depends on the \code{matching_type} and the underlying matching method used.
}
\description{
This function performs matching on a dataset based on the specified \code{matching_type}.
It supports different matching methods and configurations, such as fixed-radius or
k-nearest-neighbor methods.
}
\details{
The function abstracts the matching process, allowing for different types of matching algorithms
to be applied to the same dataset. The input \code{matching_type} determines the matching method,
metric, and additional parameters.
}
\examples{
# Example for 'maximum_fixed_scm' matching type
test_df <- data.frame(
  Z = c(1, 0, 0, 0, 1),
  X = c(0, 0.5, 0.8, 3, 1.6)
)
scaling <- 1
result <- get_matches(
  matching_type = "maximum_fixed_scm",
  df_dgp = test_df,
  scaling = scaling
)

# Example for 'euclidean_knn' matching type
test_df <- data.frame(
  Z = c(1, 0, 0, 0, 1),
  X1 = c(0, 0.5, 0.8, 3, 1.6),
  X2 = c(0, 0, 0, 0, 0)
)
result <- get_matches(
  matching_type = "euclidean_knn",
  df_dgp = test_df,
  scaling = scaling
)

}
