% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{get_plug_in_SE}
\alias{get_plug_in_SE}
\title{Calculate the standard error estimate}
\usage{
get_plug_in_SE(N_T, ESS_C, sigma_hat)
}
\arguments{
\item{N_T}{number of treated}

\item{ESS_C}{effective size of controls}

\item{sigma_hat}{The estimated error standard deviation}
}
\value{
The plug-in standard error estimate of the matching estimator
}
\description{
Calculate the standard error estimate
}
