% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{get_radius_size}
\alias{get_radius_size}
\title{get the radius size for each treated unit}
\usage{
get_radius_size(dm, rad_method, caliper)
}
\arguments{
\item{dm}{data}

\item{rad_method}{Method for radius}

\item{caliper}{Caliper size}
}
\description{
get the radius size for each treated unit
}
