% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{get_se_AE}
\alias{get_se_AE}
\title{Title}
\usage{
get_se_AE(
  matches,
  outcome = "Y",
  treatment = "Z",
  var_weight_type = "ess_units"
)
}
\arguments{
\item{matches}{The CSM match object, an R S3 object}

\item{outcome}{Name of the outcome variable (default "Y")}

\item{treatment}{Name of the treatment variable (default "Z")}

\item{var_weight_type}{The way that cluster variances are averaged
"num_units": weight by number of units in the subclass
"ess_units": weight effective size of units in the subclass
"uniform: weight each cluster equally}
}
\description{
Title
}
