% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_SL_fit}
\alias{get_SL_fit}
\title{SuperLearner fitting}
\usage{
get_SL_fit(df_to_fit, X_names, Y_name, SL.library)
}
\arguments{
\item{df_to_fit}{The data frame for whom we fit the SuperLearner.
Should contain both X and Y}

\item{X_names}{A vector of strings, specifying column names of covariates}

\item{Y_name}{A string variable specifying}

\item{SL.library}{A string to tell SuperLearner which learner to use}
}
\value{
A SuperLearner object
}
\description{
SuperLearner fitting
}
\examples{
get_SL_fit <- function(df_to_fit=df_to_fit_i,X_names=c("X1","X2"),Y_name="Y", SL.library="SL.lm")
}
