% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{impact_curve}
\alias{impact_curve}
\title{Impact curve plot}
\usage{
impact_curve(scm, outcome)
}
\arguments{
\item{scm}{A csm_matches object}

\item{outcome}{The name of the outcome variable to plot.}
}
\value{
A ggplot object showing the impact curve. Also has an
attribute "table" with the underlying data used to create the
plot.
}
\description{
Plot the relationship between maximum distance in matched set and
the outcome difference within matched sets.
}
