% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_matches_object.R
\name{is.csm_matches}
\alias{is.csm_matches}
\title{Check if object is a csm_matches object}
\usage{
is.csm_matches(x)
}
\arguments{
\item{x}{object to check}
}
\value{
logical
}
\description{
Check if object is a csm_matches object
}
