% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{love_plot}
\alias{love_plot}
\title{Love plot of covariate balance}
\usage{
love_plot(res, covs, covs_names = NULL, B = NA)
}
\description{
Make a ggplot love plot of covariate balance for each covariate
passed.
}
