% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_matches_object.R
\name{params}
\alias{params}
\title{Return the parameters of the method}
\usage{
params(csm)
}
\arguments{
\item{csm}{A csm_matches object}
}
\value{
A list of the settings used in the matching call
}
\description{
Return the parameters of the method
}
