% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_matches_object.R
\name{print.csm_matches}
\alias{print.csm_matches}
\title{Print method for csm_matches object}
\usage{
\method{print}{csm_matches}(x, ...)
}
\arguments{
\item{x}{object to print}
}
\description{
Print method for csm_matches object
}
