% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{scale_covs}
\alias{scale_covs}
\title{Scale covariates by scaling factor specified by the "scaling" variable}
\usage{
scale_covs(covs, scaling)
}
\arguments{
\item{covs}{A dataframe of covariates}

\item{scaling}{A vector of length = ncol(covs), or a scalar}
}
\value{
Scaled matrix of same dimension as covs
}
\description{
Note this converts a dataframe to a matrix.
}
