% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_plot.R
\name{sensitivity_plot}
\alias{sensitivity_plot}
\title{Make sensitivity plot of changing caliper}
\usage{
sensitivity_plot(
  res,
  data,
  min_cal = 0.05,
  max_cal = 5,
  R = 30,
  return_table = FALSE
)
}
\arguments{
\item{res}{A csm_matches object, used to get the settings for
matching}

\item{data}{The data the csm_matches object was fit to. (One could
put in alternate data here, if the covariates aligned.)}

\item{min_cal}{Minimum caliper to try}

\item{max_cal}{Maximum caliper to try}

\item{R}{Number of calipers to try between min and max}

\item{return_table}{If TRUE, return a table of results instead of a
plot}
}
\value{
A ggplot object showing the sensitivity plot.  It has an
attribute of "table" that holds the data used to make the plot.
}
\description{
Run caliper from a min to max value, generating matches at each
step. Then make a plot of the ATT and confidence interval for each
point on the plot. Also can return table of results if
'return_table' is set to TRUE.
}
