% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_matches_object.R
\name{treatment_table}
\alias{treatment_table}
\title{Get the table of treated units}
\usage{
treatment_table(csm, id = NULL, threshold = NULL, bad = FALSE)
}
\arguments{
\item{csm}{A csm_matches object}

\item{id}{Optional list of ids for treated units}

\item{threshold}{Optional distance threshold for treated units;
return only units below this threshold unless "bad" is set to
TRUE.}

\item{bad}{If TRUE, return only treated units above the threshold.
If bad is TRUE and threshold is NULL, return treated units above
the set caliper.}
}
\value{
Dataframe of treated units
}
\description{
Get the table of treated units
}
