% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_matches_object.R
\name{unmatched_units}
\alias{unmatched_units}
\title{Obtain table of treated units that were not matched}
\usage{
unmatched_units(csm)
}
\value{
dataframe, one row per treated unit.
}
\description{
Obtain table of treated units that were not matched
}
