% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_matches_object.R
\name{update_matches}
\alias{update_matches}
\title{Update a matching call to change some parameters}
\usage{
update_matches(res, data, ...)
}
\arguments{
\item{res}{A csm_matches object}

\item{...}{Parameters to change in the matching call}
}
\value{
A new csm_matches object with updated parameters
}
\description{
Update a matching call to change some parameters
}
\examples{
# Generate example data
set.seed(4044440)
dat <- gen_one_toy(nt = 5)

# Perform matching
mtch <- get_cal_matches(dat,
                        metric = "maximum",
                        scaling = c(1/0.2, 1/0.2),
                        caliper = 1,
                        rad_method = "adaptive",
                        est_method = "scm")

# View matching results
mtch

update_matches( mtch, caliper = 0.05, rad_method = "fixed" )

}
