% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{AggregationMethods}
\alias{AggregationMethods}
\alias{agg_sc_units}
\alias{agg_sc_units,}
\alias{agg_co_units,}
\alias{agg_avg_units}
\alias{agg_co_units}
\title{aggregation methods for matched data}
\usage{
agg_sc_units(scweights)

agg_co_units(scweights)

agg_avg_units(scweights)
}
\arguments{
\item{scweights}{scm_matches object or list of sc weights (tibbles}
}
\value{
data.frame with tx and corresponding control units.
}
\description{
agg_sc_units weights to make pairs of tx and synthetic control
units. Generally aggregate by cluster defined by treated unit,
calculating the weighted average of the control units in the
cluster for all variables that start with "X" or "Y" in the
dataframes.
}
\details{
agg_co_units Aggregates across the controls, grouping repeated
controls used by multiple treated units

agg_avg_units does simple average of controls within groups (e.g.,
like CEM would do).
}
