% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot_CSM}
\alias{boot_CSM}
\title{Title}
\usage{
boot_CSM(
  dgp_name,
  att0,
  I = 100,
  B = 250,
  mu_model = "linear",
  boot_mtd = "Bayesian",
  n_split = 1,
  kang_true = F,
  toy_ctr_dist = 0.5
)
}
\arguments{
\item{dgp_name}{Name of the DGP}

\item{kang_true}{}
}
\description{
Title
}
\examples{
boot_otsu_to_test <-
boot_CSM(dgp_name="otsu", att0=T,I=100, B=100,mu_model="linear", boot_mtd="Bayesian",n_split=2)
}
