% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{coerce_covs}
\alias{coerce_covs}
\title{Coerce non-numeric factor columns to integer values}
\usage{
coerce_covs(covs)
}
\arguments{
\item{covs}{2-d dataframe of N x p, where N is number of data,
p is dimension of features}
}
\description{
Coerce non-numeric factor columns to integer values
}
