% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{est_weights}
\alias{est_weights}
\title{Generate weights for list of matched sets}
\usage{
est_weights(
  df,
  matched_gps,
  dist_scaling,
  est_method = c("scm", "average"),
  metric = c("maximum", "euclidean", "manhattan")
)
}
\arguments{
\item{df}{full dataframe}

\item{matched_gps}{list of matched sets: tx unit in first row}

\item{dist_scaling}{tibble with scaling for each covariate}

\item{est_method}{"scm" or "average"}
}
\value{
list of matched sets, with 'weights'
}
\description{
Generate weights for list of matched sets
}
