% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_matches_object.R
\name{feasible_units}
\alias{feasible_units}
\title{Return table of feasible treated units}
\usage{
feasible_units(csm)
}
\value{
Dataframe of all the treated units (not controls) that were
matched within a caliper.
}
\description{
Return those treated units with controls within the set caliper.
}
