% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{gen_df_adv}
\alias{gen_df_adv}
\title{Generate toy data}
\usage{
gen_df_adv(
  nc,
  nt,
  f0_sd = 0.1,
  f0_fun = function(X1, X2) {
     1
 },
  tx_effect_fun = function(X1, X2) {
     1
 },
  ctr_dist = 0.5,
  prop_nc_unif = 1/3
)
}
\arguments{
\item{nc}{(something)}

\item{nt}{(something)}

\item{f0_sd}{(something)}

\item{f0_fun}{(something)}

\item{tx_effect_fun}{(something)}

\item{ctr_dist}{(something)}

\item{prop_nc_unif}{(something)}
}
\value{
A tibble containing a toy dataset
}
\description{
Generate toy data
}
