% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{gen_dm}
\alias{gen_dm}
\title{Generate (#tx) by (#co) distance matrix}
\usage{
gen_dm(
  df,
  covs = starts_with("X"),
  treatment = Z,
  scaling = 1,
  metric = c("maximum", "euclidean", "manhattan")
)
}
\arguments{
\item{df}{dataframe}

\item{covs}{covariates; default all "X" variables}

\item{treatment}{treatment variable; default "Z"}

\item{scaling}{vector of scaling constants for covariates}

\item{metric}{distance metric to use}
}
\value{
(#tx) by (#co) distance matrix
}
\description{
Generate (#tx) by (#co) distance matrix
}
