% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{gen_matches}
\alias{gen_matches}
\title{Generate matches for each treatment over controls
using specified covariate names,
distance metric specified by the scaling parameter
and the method of metric.
Finally, there is a selection of the units}
\usage{
gen_matches(
  df,
  covs = starts_with("X"),
  treatment = "Z",
  scaling = 1,
  metric = "maximum",
  caliper = 1,
  rad_method = c("adaptive", "1nn", "fixed"),
  ...
)
}
\arguments{
\item{df}{A df containing at least covs and treatment}

\item{metric}{Character specifying the distance metric.
One of "maximum", "euclidean", "manhattan"}

\item{caliper}{Caliper on the scaled distance. Usually set to 1, and control matches
using the scaling paramter}

\item{...}{}
}
\description{
Note: we allow controls to be repeatedly used
}
