% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{gen_one_toy}
\alias{gen_one_toy}
\title{Generate a toy dataset with a single treatment effect}
\usage{
gen_one_toy(
  nc = 500,
  nt = 100,
  ctr_dist = 0.5,
  prop_nc_unif = 1/3,
  f0_sd = 0.5
)
}
\arguments{
\item{ctr_dist}{Distance between the two control clusters}
}
\value{
A tibble with the toy dataset
}
\description{
Generate a toy dataset with a single treatment effect
}
