% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{get_ATT_estimate}
\alias{get_ATT_estimate}
\title{Estimate ATT}
\usage{
get_ATT_estimate(scmatch, outcome = "Y")
}
\description{
Calculate ATT and associated standard error using the weighting
method from paper with locally estimated residual variation.
}
