% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_cal_matches}
\alias{get_cal_matches}
\title{Main adaptive radius matching function}
\usage{
get_cal_matches(
  df,
  covs = starts_with("X"),
  treatment = "Z",
  metric = c("maximum", "euclidean", "manhattan"),
  caliper = 1,
  rad_method = c("adaptive", "fixed", "1nn"),
  est_method = c("scm", "scm_extrap", "average"),
  return = c("sc_units", "agg_co_units", "all"),
  dist_scaling = df \%>\% summarize(across(starts_with("X"), function(x) {
     if
    (is.numeric(x)) 
         1/sd(x)
     else 1000
 })),
  ...
)
}
\arguments{
\item{df}{The data frame to be matched}

\item{metric}{A string specifying the distance metric}

\item{caliper}{A numeric specifying the caliper size}

\item{rad_method}{adaptive caliper, fixed caliper, only 1nn caliper}

\item{est_method}{A string specifying the estimation method}

\item{return}{A string specifying what to return}

\item{dist_scaling}{A vector of scaling constants for covariates}

\item{...}{Additional arguments}
}
\value{
df with a bunch of attributes.
}
\description{
Main adaptive radius matching function
}
