% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{get_radius_size}
\alias{get_radius_size}
\title{get the radius size for each treated unit}
\usage{
get_radius_size(
  dm,
  rad_method = c("adaptive", "targeted", "fixed", "1nn", "knn", "knn-capped"),
  caliper,
  k = 1
)
}
\arguments{
\item{dm}{Distance metric, rows are treated units, columns are
control units}

\item{rad_method}{Method for determining radius size. Options are:
"adaptive" -- grow caliper to capture at least k units,
"targeted" -- shrink caliper to target k units, but also grow so
there is at least 1 unit, "fixed" -- use caliper as is, "1nn" or
"knn" -- set caliper to kth nearest neighbor, "knn-capped" --
distance to k-th nearest neighbor but capped at caliper.}

\item{caliper}{Caliper size}

\item{k}{For knn method, number of nearest neighbors}
}
\description{
get the radius size for each treated unit
}
