% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{get_radius_size}
\alias{get_radius_size}
\title{get the radius size for each treated unit}
\usage{
get_radius_size(
  dm,
  rad_method = c("adaptive", "fixed", "1nn", "knn", "knn-capped"),
  caliper,
  k = 1
)
}
\arguments{
\item{dm}{Distance metric, rows are treated units, columns are control units}

\item{rad_method}{Method for determining radius size}

\item{caliper}{Caliper size}

\item{k}{For knn method, number of nearest neighbors}
}
\description{
get the radius size for each treated unit
}
