% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_matches_object.R
\name{result_table}
\alias{result_table}
\title{Get table of aggregated results, with rows for synthetic units if
that was asked for}
\usage{
result_table(csm, feasible_only = FALSE)
}
\value{
dataframe
}
\description{
Get table of aggregated results, with rows for synthetic units if
that was asked for
}
