% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_matches_object.R
\name{result_table}
\alias{result_table}
\title{Obtain table of aggregated results}
\usage{
result_table(
  csm,
  return = c("all", "sc_units", "agg_co_units"),
  feasible_only = FALSE,
  nonzero_weight_only = FALSE
)
}
\arguments{
\item{return}{How to aggregate units, if at all, in making the
result table.  Possible values: "sc_units" (the synthetic control
units), "agg_co_units" (the unique control units, with total
weight across all their matches), or "all" (control units will be
repeated if matched multiply). Defaults to "all".}

\item{feasible_only}{TRUE means only return units which were
matched without expanding the caliper.}
}
\value{
dataframe of the treatment and control units.  This
dataframe can be analyzed as an as-if experimental dataset.
}
\description{
This method will give the controls as synthetic controls
(sc_units), the individual controls with repeat rows if controls
were used for different treated units (all), or aggregated controls
with only one row per control unit (agg_co_units).
}
