% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_matches_object.R
\name{summary.csm_matches}
\alias{summary.csm_matches}
\title{Summary method for csm_matches object}
\usage{
\method{summary}{csm_matches}(x, outcome = NULL, ...)
}
\arguments{
\item{x}{object to summarize}

\item{...}{Extra arguments (currently ignored).}
}
\description{
Summary method for csm_matches object
}
